refactor: deprecated 문제로 jwts 버전 다운그레이드 및 코드 리팩토링
