+++ src/main/java/ject/petfit/PetfitApplication.java
@@ -2,6 +2,7 @@ package ject.petfit;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 public class PetfitApplication {
