+++ src/main/java/ject/petfit/domain/user/controller/KakaoAuthUserController.java
@@ -10,8 +10,8 @@ import ject.petfit.domain.user.dto.response.AuthUserResponseDTO;
 import ject.petfit.global.jwt.util.JwtUtil;
 import ject.petfit.global.jwt.refreshtoken.RefreshToken;
 import ject.petfit.global.jwt.refreshtoken.RefreshTokenService;
-import ject.petfit.global.jwt.token.RefreshTokenRequestDTO;
-import ject.petfit.global.jwt.token.TokenResponseDTO;
+import ject.petfit.global.jwt.dto.RefreshTokenRequestDTO;
+import ject.petfit.global.jwt.dto.TokenResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
