+++ src/main/java/ject/petfit/global/jwt/dto/RefreshTokenRequestDTO.java
@@ -1,4 +1,4 @@
-package ject.petfit.global.jwt.token;
+package ject.petfit.global.jwt.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,6 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class TokenResponseDTO {
-    private String accessToken;
+public class RefreshTokenRequestDTO {
     private String refreshToken;
 }
